+++ src/abstract-reporter.js
