+++ src/errors/base-error.js
