+++ src/errors/worker-error.js
