+++ src/errors/worker-stop-error.js
