+++ src/rollbar-reporter.js
